support NO_NEED_CONVERT(just change package)
